api/requirements.txt
fastapi==0.111.0
pydantic==2.7.1
mysql-connector-python==8.4.0
gunicorn==22.0.0